=map javascript

+++ Map/map.php
@@ -86,10 +86,10 @@
 							console.log(msg.feed.entry[0].gsx$maternalmortalityratioper100000livebirths.$t);
 							console.log(msg.feed.entry[0].gsx$neonatalmortalityrateper1000livebirths.$t);
 							console.log(friend_count);
-						console.log("27.5 "/1000);
+							console.log("27.5 "/1000);
 							
-							var mummyDeathsPer100000 = (Double)msg.feed.entry[0].gsx$maternalmortalityratioper100000livebirths.$t;
-							var babyDeathsPer1000 = (Double)msg.feed.entry[0].gsx$neonatalmortalityrateper1000livebirths.$t
+							var mummyDeathsPer100000 = msg.feed.entry[0].gsx$maternalmortalityratioper100000livebirths.$t;
+							var babyDeathsPer1000 = msg.feed.entry[0].gsx$neonatalmortalityrateper1000livebirths.$t
 							mummyDeaths = (mummyDeathsPer100000 / 100000) * friend_count;
 							babyDeaths = (babyDeathsPer1000 / 1000) * friend_count;
 						}
